move update stream to types
